Refactor gk-rz into the naming conventions which are used by GK and GK specific, and refactor out the core methods of interpolating on geometry into map.py in operations. Explicit geometry construction is kept general, but the GK-specific nature is kept in diagnostics. Examples are updated. CLI names are still gk_rz. Script pg.gk.rz() is a short hand for chaining a few core commands together, including constructing the geometry object and the projection.
Entire-Checkpoint: 01M26HA71FS8K7HRX1BNFMXXQ5
