+++ docs/source/physical-rz.rst
@@ -9,6 +9,13 @@ The Python and CLI outputs below are both generated and compared when this websi
 Computational coordinates are useful for analysis, but a poloidal view places
 the field in its physical geometry. Keep the companion
 ``rt_gk_tcv_nt_iwl_3x2v_p1-geo_int_mapc2p.gkyl`` beside the electron density
-file so the operation can resolve it by the simulation prefix.
+file so ``pg.gk.rz`` can resolve it by the simulation prefix. This diagnostic
+loads Gkeyll geometry and composes equation-independent coordinate operations.
+
+For explicit geometry, construct ``pg.Geometry`` from coordinate arrays, build
+``pg.resolve_rz_projection(data, geometry)``, and apply it with
+``data.map_to_rz(projection=projection)``. The projection can be reused for
+fields on the same computational grid. Three-dimensional reconstruction
+assumes periodic field-aligned coordinates and twist-and-shift boundaries.
 
 .. include:: _pairs/05_gk_rz.inc
