+++ src/postgkyl/diagnostics/__init__.py
@@ -5,9 +5,8 @@ Folds together the old ``models`` (array math) and ``operations`` physics-verb
 here take loaded ``GData``/``GDataState`` (one or several) plus physical
 scalars as keyword-only options, and return a ``GDataState`` (via
 ``_result``) or, in later layers, a ``Figure``. Equation-blind core verbs
-stay in flat ``operations`` modules. Domain-specific transformations live in
-operation subpackages (for example ``operations.gyrokinetics``); this layer
-is reserved for code that knows what field components physically mean.
+stay in flat ``operations`` modules. Model-specific auxiliary discovery and
+interpretation live beside their diagnostic compositions.
 
 The four public packages mirror Gkeyll's model families: ``gk``, ``vm``,
 ``pkpm``, and ``mom``. The equation-blind ``discovery`` module
@@ -150,11 +149,7 @@ for _name in (
     "is_geo_mapc2p",
     "multib_tag",
     "nodes_to_RZ",
-    "map_to_rz",
     "resolve_geometry",
-    "resolve_rz_projection",
-    "extract_flux_surface",
-    "resolve_flux_surface_grid",
 ):
   _function = getattr(gk, _name)
   if command_spec(_function) is None and hidden_spec(_function) is None:
